+++ templates/includes/_header.html
@@ -48,7 +48,7 @@
                             {% endif %}
                             <span style="padding:0.5em">|</span>
                             <li class="nav-item">
-                                <a class="nav-link" href="{% url 'pages:help_home' %}">🛟 {% translate "Help" %}</a>
+                                <a class="nav-link" href="{% url 'pages:help' %}">🛟 {% translate "Help" %}</a>
                             </li>
                         {% endif %}
                     </ul>
