+++ templates/includes/_header_notice_beta.html
@@ -2,6 +2,6 @@
 
 <div id="header-notice-beta" class="alert alert-info font-weight-bold text-center p-2 m-0">
     <p class="mb-0">
-        {% translate "Website under construction. Feedback is gladly welcome (bugs, improvements...)" %} 🙏🏻👉 <a href="{% url 'pages:help_home' %}#contact">{% translate "contact us" %}</a>
+        {% translate "Website under construction. Feedback is gladly welcome (bugs, improvements...)" %} 🙏🏻👉 <a href="{% url 'pages:help' %}#contact">{% translate "contact us" %}</a>
     </p>
 </div>
