+++ templates/users/administrator_list.html
@@ -11,7 +11,7 @@
             <nav style="--bs-breadcrumb-divider: '>';" aria-label="breadcrumb">
                 <ol class="breadcrumb">
                     <li class="breadcrumb-item"><a href="{% url 'pages:home' %}">{% translate "Home" %}</a></li>
-                    <li class="breadcrumb-item"><a href="{% url 'pages:help_home' %}">{% translate "Help" %}</a></li>
+                    <li class="breadcrumb-item"><a href="{% url 'users:home' %}">{% translate "Users" %}</a></li>
                     <li class="breadcrumb-item active" aria-current="page">{% translate "Administrator list" %}</li>
                 </ol>
             </nav>
