+++ templates/users/home.html
@@ -39,6 +39,7 @@
             <div class="card">
                 <div class="card-body">
                     <h5 class="card-title">{% translate "Administrators" %} <small class="text-muted">{{ administrator_count }}</small></h5>
+                    <a class="btn btn-primary" href="{% url 'users:administrator_list' %}">{% translate "View" %}</a>
                 </div>
             </div>
         </div>
